remove minifier, fix bundler when no file exists

+++ bundle.mjs
@@ -1,8 +1,6 @@
 /* Simple bundler that combines the html, css, and JS files */
 import fs, { unlink } from "fs";
-import { minify } from "html-minifier";
 import path from "path";
-import { exit } from "process";
 import { fileURLToPath } from "url";
 
 const html = "xml-anonymizer.html";
@@ -18,39 +16,40 @@ const jsFilePath = path.join(srcPath, script);
 const cssFilePath = path.join(srcPath, css);
 const outputPath = path.join(__dirname, "/dist/anonymizer.html");
 
-console.log(`Deleting ${outputPath}...`);
-unlink(outputPath, (err) => {
-  if (err) console.error("error,", err);
-  else {
-    console.log(
-      `\x1b[32m${outputPath}\x1b[0m deleted successfully, building...`
-    );
-    console.log("bundling files...");
-    /* Read in individual files */
-    const htmlContent = fs.readFileSync(htmlFilePath, "utf8");
-    const jsContent = fs.readFileSync(jsFilePath, "utf8");
-    const cssContent = fs.readFileSync(cssFilePath, "utf8");
+if (fs.existsSync(outputPath)) {
+  console.log(`Deleting ${outputPath}...`);
+  unlink(outputPath, (err) => {
+    if (err) console.error("error,", err);
+    else {
+      console.log(
+        `\x1b[32m${outputPath}\x1b[0m deleted successfully, building...`,
+      );
+      bundleFiles();
+    }
+  });
+} else {
+  bundleFiles();
+}
 
-    /* Replace External Script with inline JS */
-    let bundledHTML = htmlContent.replace(
-      /<script\s+src=["']xml.js["']><\/script>/,
-      `<script>${jsContent}</script>`
-    );
+function bundleFiles() {
+  console.log("bundling files...");
+  /* Read in individual files */
+  const htmlContent = fs.readFileSync(htmlFilePath, "utf8");
+  const jsContent = fs.readFileSync(jsFilePath, "utf8");
+  const cssContent = fs.readFileSync(cssFilePath, "utf8");
 
-    /* Replace External Stylesheet with inline CSS */
-    bundledHTML = bundledHTML.replace(
-      /<link\s+rel=["']stylesheet["']\s+href=["']styles\.css["']\s*\/>/,
-      `<style>${cssContent}</style>`
-    );
+  /* Replace External Script with inline JS */
+  let bundledHTML = htmlContent.replace(
+    /<script\s+src=["']xml.js["']><\/script>/,
+    `<script>${jsContent}</script>`,
+  );
 
-    const minifiedHtml = minify(bundledHTML, {
-      collapseWhitespace: true,
-      removeComments: true,
-      minifyCSS: true,
-      minifyJS: true,
-    });
+  /* Replace External Stylesheet with inline CSS */
+  bundledHTML = bundledHTML.replace(
+    /<link\s+rel=["']stylesheet["']\s+href=["']styles\.css["']\s*\/>/,
+    `<style>${cssContent}</style>`,
+  );
 
-    fs.writeFileSync(outputPath, minifiedHtml, "utf8");
-    console.log(`Bundled file created at: \x1b[32m${outputPath}`);
-  }
-});
+  fs.writeFileSync(outputPath, bundledHTML, "utf8");
+  console.log(`Bundled file created at: \x1b[32m${outputPath}`);
+}
